commit ajustes atualizar projeto com indicadores avulsos encaminhados

+++ src/main/java/br/gov/es/siscap/dto/ProjetoIndicadorAvulsoDto.java
@@ -8,7 +8,7 @@ public record ProjetoIndicadorAvulsoDto(
 	
 	Long id,
 
-	Long idIndicadorAvulso,
+	Integer idIndicadorAvulso,
 
 	@Valid
 	IndicadorAvulsoDto indicadorAvulso,
